[Refactor] transformStatsJSONtoArray()を抽出
